+++ Tests/Unit/Import/Fixtures/TestingDomDocumentConverter.php
@@ -1,12 +1,14 @@
 <?php
 
+namespace OliverKlee\Realty\Tests\Unit\Import\Fixtures;
+
 /**
  * This is merely a class used for unit tests. Don't use it for any other purpose.
  *
  * @author Saskia Metzler <saskia@merlin.owl.de>
  * @author Oliver Klee <typo3-coding@oliverklee.de>
  */
-final class tx_realty_domDocumentConverterChild extends \tx_realty_domDocumentConverter
+final class TestingDomDocumentConverter extends \tx_realty_domDocumentConverter
 {
     /**
      * Adds a new element $value to the array $arrayExpand, using the key $key.
@@ -37,8 +39,7 @@ final class tx_realty_domDocumentConverterChild extends \tx_realty_domDocumentCo
     /**
      * Creates an array of document records for one realty record.
      *
-     * @return array[]
-     *         document records, will be empty if there are none
+     * @return array[] document records, will be empty if there are none
      */
     public function importDocuments()
     {
@@ -54,8 +55,7 @@ final class tx_realty_domDocumentConverterChild extends \tx_realty_domDocumentCo
      * @param string $nameOfChild name of child, must not be empty
      * @param string $nameOfGrandchild name of grandchild, must not be empty
      *
-     * @return DOMNode first grandchild with this name, NULL if it does not
-     *                 exist
+     * @return \DOMNode|null first grandchild with this name
      */
     public function findFirstGrandchild($nameOfChild, $nameOfGrandchild)
     {
@@ -66,7 +66,7 @@ final class tx_realty_domDocumentConverterChild extends \tx_realty_domDocumentCo
      * Fetches an attribute from a given node and returns name/value pairs as an
      * array. If there are no attributes, the returned array will be empty.
      *
-     * @param DOMNode|DOMNodeList|null $nodeWithAttributes node from where to fetch the attribute, may be NULL
+     * @param \DOMNode|\DOMNodeList|null $nodeWithAttributes node from where to fetch the attribute, may be NULL
      *
      * @return string[] attributes and attribute values, empty if there are no attributes
      */
@@ -78,23 +78,23 @@ final class tx_realty_domDocumentConverterChild extends \tx_realty_domDocumentCo
     /**
      * Substitutes XML namespaces from a node name and returns the name.
      *
-     * @param DOMNode|null $domNode node, may be NULL
+     * @param \DOMNode|null $domNode node, may be NULL
      *
      * @return string node name without namespaces, may be empty
      */
-    public function getNodeName(DOMNode $domNode = null)
+    public function getNodeName(\DOMNode $domNode = null)
     {
         return parent::getNodeName($domNode);
     }
 
     /**
-     * Loads the raw data from a DOMDocument.
+     * Loads the raw data from a \DOMDocument.
      *
-     * @param DOMDocument $rawRealtyData raw data to load, must not be NULL
+     * @param \DOMDocument $rawRealtyData raw data to load, must not be NULL
      *
      * @return void
      */
-    public function setRawRealtyData(DOMDocument $rawRealtyData)
+    public function setRawRealtyData(\DOMDocument $rawRealtyData)
     {
         parent::setRawRealtyData($rawRealtyData);
     }